// src/api/generated/models/createRoleDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { CreateRoleDtoType } from './createRoleDtoType';

export interface CreateRoleDto {
  code: string;
  name: string;
  description?: string;
  type: CreateRoleDtoType;
  /**
     * @minimum 1
     * @maximum 99
     */
  level: number;
  guardName?: string;
}